src/apps/nfc/action/controllers/get-actions.controller.ts
import { Controller, Get, Body } from '@nestjs/common';
import { ApiTags, ApiOkResponse, ApiOperation, ApiBody, ApiQuery } from '@nestjs/swagger';
import { ActionDto } from './../dto/action.dto';

// @hades
import { IQueryBus } from '@hades/shared/domain/bus/query-bus';
import { QueryStatementInput } from '@hades/shared/domain/persistence/sql-statement-input';
import { GetActionsQuery } from '@hades/nfc/action/application/get/get-actions.query';

@ApiTags('[nfc] action')
@Controller('nfc/actions')
export class GetActionsController 
{
    constructor(
        private readonly queryBus: IQueryBus
    ) {}

    @Get()
    @ApiOperation({ summary: 'Find actions according to query' })
    @ApiOkResponse({ description: 'The records has been found successfully.', type: [ActionDto] })
    @ApiBody({ type: [QueryStatementInput] })
    @ApiQuery({ name: 'query', type: [QueryStatementInput] })
    async main(@Body('query') queryStatements: QueryStatementInput[])
    {
        return await this.queryBus.ask(new GetActionsQuery(queryStatements));   
    }
}